refactor: Update getDefaultLanguageCode function to handle desktop and web environments
The code changes in this commit update the getDefaultLanguageCode function in the Language/Utils module. It introduces conditional logic to handle different environments: desktop and web. In the desktop environment, the function fetches the language code from an API based on the organization URL and other parameters. In the web environment, the function calls the getDefaultLanguageCodeWeb function. This change improves the functionality of the getDefaultLanguageCode function and ensures the correct language code is retrieved based on the environment.

Update getDefaultLanguageCode function to handle desktop and web environments.
